+++ src/Admin/Controls/ProductGridFiltersFactory.php
@@ -235,33 +235,33 @@ class ProductGridFiltersFactory
 			}, '', 'displayAmount', null, $displayAmounts, ['placeholder' => '- Dostupnost -']);
 		}
 
-		$grid->addFilterDataSelect(function (ICollection $source, $value): void {
-			$source->where('this.hidden', (bool) $value);
-		}, '', 'hidden', null, ['1' => 'Skryté', '0' => 'Viditelné'])->setPrompt('- Viditelnost -');
-
-		$grid->addFilterDataSelect(function (ICollection $source, $value): void {
-			$source->where('this.recommended', (bool) $value);
-		}, '', 'recommended', null, ['1' => 'Doporučené', '0' => 'Normální'])->setPrompt('- Doporučené -');
-
-		$grid->addFilterDataSelect(function (ICollection $source, $value): void {
-			$source->where('this.unavailable', (bool) $value);
-		}, '', 'unavailable', null, ['1' => 'Neprodejné', '0' => 'Prodejné'])->setPrompt('- Prodejnost -');
+//		$grid->addFilterDataSelect(function (ICollection $source, $value): void {
+//			$source->where('visibilityListItem.hidden', (bool) $value);
+//		}, '', 'hidden', null, ['1' => 'Skryté', '0' => 'Viditelné'])->setPrompt('- Viditelnost -');
+//
+//		$grid->addFilterDataSelect(function (ICollection $source, $value): void {
+//			$source->where('this.recommended', (bool) $value);
+//		}, '', 'recommended', null, ['1' => 'Doporučené', '0' => 'Normální'])->setPrompt('- Doporučené -');
+//
+//		$grid->addFilterDataSelect(function (ICollection $source, $value): void {
+//			$source->where('this.unavailable', (bool) $value);
+//		}, '', 'unavailable', null, ['1' => 'Neprodejné', '0' => 'Prodejné'])->setPrompt('- Prodejnost -');
 
 		$grid->addFilterDataSelect(function (ICollection $source, $value): void {
 			if ($value === 'green') {
 				$source->setGroupBy(
 					['this.uuid'],
-					'this.hidden = "0" AND this.unavailable = "0" AND COUNT(DISTINCT price.uuid) > 0 AND COUNT(DISTINCT nxnCategory.fk_category) > 0 AND pricelistActive = "1"',
+					'visibilityListItem.hidden = "0" AND this.unavailable = "0" AND COUNT(DISTINCT price.uuid) > 0 AND COUNT(DISTINCT nxnCategory.fk_category) > 0 AND pricelistActive = "1"',
 				);
 			} elseif ($value === 'orange') {
 				$source->setGroupBy(
 					['this.uuid'],
-					'this.hidden = "0" AND this.unavailable = "0" AND COUNT(DISTINCT price.uuid) > 0 AND COUNT(DISTINCT nxnCategory.fk_category) = 0 AND pricelistActive = "1"',
+					'visibilityListItem.hidden = "0" AND this.unavailable = "0" AND COUNT(DISTINCT price.uuid) > 0 AND COUNT(DISTINCT nxnCategory.fk_category) = 0 AND pricelistActive = "1"',
 				);
 			} else {
 				$source->setGroupBy(
 					['this.uuid'],
-					'this.hidden = "1" OR this.unavailable = "1" OR COUNT(DISTINCT price.uuid) = 0 OR COUNT(DISTINCT nxnCategory.fk_category) = 0 OR pricelistActive = "0"',
+					'visibilityListItem.hidden = "1" OR this.unavailable = "1" OR COUNT(DISTINCT price.uuid) = 0 OR COUNT(DISTINCT nxnCategory.fk_category) = 0 OR pricelistActive = "0"',
 				);
 			}
 		}, '', 'show', null, ['green' => 'Viditelné', 'orange' => 'Viditelné: bez kategorie', 'red' => 'Neviditelné'])->setPrompt('- Viditelnost v eshopu -');