+++ src/Admin/VisibilityListPresenter.php
@@ -5,7 +5,10 @@ namespace Eshop\Admin;
 use Admin\Controls\AdminForm;
 use Admin\Controls\AdminGrid;
 use Eshop\BackendPresenter;
+use Eshop\DB\CategoryRepository;
+use Eshop\DB\CategoryTypeRepository;
 use Eshop\DB\Product;
+use Eshop\DB\ProductRepository;
 use Eshop\DB\VisibilityList;
 use Eshop\DB\VisibilityListItem;
 use Eshop\DB\VisibilityListRepository;
@@ -13,7 +16,9 @@ use Forms\Form;
 use Grid\Datagrid;
 use Nette\Application\Attributes\Persistent;
 use Nette\DI\Attributes\Inject;
+use Nette\Utils\Strings;
 use StORM\Collection;
+use StORM\ICollection;
 
 class VisibilityListPresenter extends BackendPresenter
 {
@@ -27,6 +32,15 @@ class VisibilityListPresenter extends BackendPresenter
 
 	#[Inject]
 	public \Eshop\DB\VisibilityListItemRepository $visibilityListItemRepository;
+
+	#[Inject]
+	public CategoryTypeRepository $categoryTypeRepository;
+
+	#[Inject]
+	public CategoryRepository $categoryRepository;
+
+	#[Inject]
+	public ProductRepository $productRepository;
 
 	#[Persistent]
 	public string $tab = 'lists';
@@ -81,7 +95,10 @@ class VisibilityListPresenter extends BackendPresenter
 
 		$grid->addColumnInputInteger('Priorita', 'priority', '', '', 'priority', [], true);
 
+		$grid->addColumnInputCheckbox('<i title="Doporučeno" class="far fa-thumbs-up"></i>', 'recommended', '', '', 'recommended');
 		$grid->addColumnInputCheckbox('<i title="Skryto" class="far fa-eye-slash"></i>', 'hidden', '', '', 'hidden');
+		$grid->addColumnInputCheckbox('<i title="Skryto v menu a vyhledávání" class="far fa-minus-square"></i>', 'hiddenInMenu', '', '', 'hiddenInMenu');
+		$grid->addColumnInputCheckbox('<i title="Neprodejné" class="fas fa-ban"></i>', 'unavailable', '', '', 'unavailable');
 
 		$grid->addColumnLinkDetail('itemDetail');
 		$grid->addColumnActionDelete();
@@ -97,6 +114,49 @@ class VisibilityListPresenter extends BackendPresenter
 				$source->where('this.fk_visibilityList', $value);
 			}, '', 'list', null, $categories)->setPrompt('- Seznam -');
 		}
+
+		$firstCategoryType = $this->categoryTypeRepository->many()->setOrderBy(['priority'])->first();
+
+		if ($firstCategoryType && ($categories = $this->categoryRepository->getTreeArrayForSelect(true, $firstCategoryType->getPK()))) {
+			$exactCategories = $categories;
+			$categories += ['0' => 'X - bez kategorie'];
+
+			foreach ($exactCategories as $key => $value) {
+				$categories += ['.' . $key => $value . ' (bez podkategorií)'];
+			}
+
+			$grid->addFilterDataSelect(function (Collection $source, $value): void {
+				if (\str_starts_with($value, '.')) {
+					$subSelect = $this->categoryRepository->getConnection()->rows(['eshop_product_nxn_eshop_category'])
+						->where('this.fk_product = eshop_product_nxn_eshop_category.fk_product')
+						->where('eshop_product_nxn_eshop_category.fk_category', Strings::substring($value, 1));
+
+					$source->where('EXISTS (' . $subSelect->getSql() . ')', $subSelect->getVars());
+				} else {
+					$category = $this->categoryRepository->one($value);
+
+					if (!$category && $value !== '0') {
+						$source->where('1=0');
+
+						return;
+					}
+
+					$source->filter(['category' => $value === '0' ? false : $category->path]);
+				}
+			}, '', 'category', null, $categories)->setPrompt('- Kategorie -');
+		}
+
+		$grid->addFilterDataSelect(function (ICollection $source, $value): void {
+			$source->where('this.hidden', (bool) $value);
+		}, '', 'hidden', null, ['1' => 'Skryté', '0' => 'Viditelné'])->setPrompt('- Viditelnost -');
+
+		$grid->addFilterDataSelect(function (ICollection $source, $value): void {
+			$source->where('this.recommended', (bool) $value);
+		}, '', 'recommended', null, ['1' => 'Doporučené', '0' => 'Normální'])->setPrompt('- Doporučené -');
+
+		$grid->addFilterDataSelect(function (ICollection $source, $value): void {
+			$source->where('this.unavailable', (bool) $value);
+		}, '', 'unavailable', null, ['1' => 'Neprodejné', '0' => 'Prodejné'])->setPrompt('- Prodejnost -');
 
 		$grid->addFilterButtons();
 
@@ -149,8 +209,9 @@ class VisibilityListPresenter extends BackendPresenter
 
 			$hiddenInput = $form->addCheckbox('hidden', 'Skryto');
 			$hiddenInMenuInput = $form->addCheckbox('hiddenInMenu', 'Skryto v menu a vyhledávání');
-
 			$hiddenInMenuInput->addConditionOn($hiddenInput, $form::EQUAL, false)->toggle($hiddenInMenuInput->getHtmlId() . '-toogle');
+
+			$form->addCheckbox('unavailable', 'Neprodejné')->setHtmlAttribute('data-info', 'Znemožňuje nákup produktu.');
 
 			$productInput = $form->addSelectAjax('product', 'Produkt', 'Zvolte produkt', Product::class);